Telemetry does not ned to be an initilizer
